refactor(webpack): better preload setup
No more prefetching - not currently used and not supported in Chrome in
a CSP yet. Switching to preloading. Excluding index since the only thing
that uses preloading is fireworks, which would have otherwise been
included in the index because we're not using route-based code splitting
yet. Fixes the stylesheet to not generate and preload its own chunk.
Sadly it's still in the others, but it's not worth the extra code nor
reduced file size to fix.